move python SMD code to experiments
